Add support for ruby 2.2.3
**Why this change?**

On 18-08 ruby 2.2.3 was released with a sercurity fix. I think it would be good
for IC to start supporting that version asap.

At first installation of servers with ruby 2.2.3 will be a bit slower, this is
because we did not yet build the binaries for this ruby version. Those will be
released in the next few days.

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>